pyproject.toml
[project]
name = "finlactics-phidata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.135.2",
    "groq>=1.1.2",
    "openai>=2.30.0",
    "packaging>=26.0",
    "phidata>=2.7.10",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
    "yfinance>=1.2.0",
]
